extract prices to own component and fix responsive
