+++ web/components/mint.tsx
@@ -1,100 +1,35 @@
 import React, { useContext, useState } from 'react';
 import { Helmet } from 'react-helmet';
-import { stacksNetwork as network, getRPCClient } from '@common/utils';
+import { stacksNetwork as network } from '@common/utils';
 import { useSTXAddress } from '@common/use-stx-address';
-import BN from 'bn.js';
 import {
   AnchorMode,
-  broadcastTransaction,
   callReadOnlyFunction,
   cvToJSON,
-  createStacksPrivateKey,
   standardPrincipalCV,
-  makeSTXTokenTransfer,
-  privateKeyToString,
   uintCV,
 } from '@stacks/transactions';
 import { VaultGroup } from './vault-group';
-import { getPriceInfo, getDikoAmmPrice, getUsdaPrice } from '@common/get-price';
 import { AppContext } from '@common/context';
 import { useConnect } from '@stacks/connect-react';
 import { CollateralType } from '@components/collateral-type';
 import { useEffect } from 'react';
-import { tokenList } from '@components/token-swap-list';
 import { VaultProps } from './vault';
 import { EmptyState } from './ui/empty-state';
 import { ArchiveIcon } from '@heroicons/react/outline';
 import { Placeholder } from './ui/placeholder';
 import { InformationCircleIcon } from '@heroicons/react/solid';
 import { Tooltip } from '@blockstack/ui';
+import { Prices } from './prices';
 
 export const Mint = () => {
   const address = useSTXAddress();
-  const env = process.env.REACT_APP_NETWORK_ENV || 'regtest';
   const [state, setState] = useContext(AppContext);
   const [{ vaults, collateralTypes }, _x] = useContext(AppContext);
   const { doContractCall } = useConnect();
   const contractAddress = process.env.REACT_APP_CONTRACT_ADDRESS || '';
-  const [stxPrice, setStxPrice] = useState(0.0);
-  const [dikoPrice, setDikoPrice] = useState(0.0);
-  const [xbtcPrice, setXbtcPrice] = useState(0.0);
-  const [usdaPrice, setUsdaPrice] = useState(1.0);
-  const [stxBlockUpdate, setStxBlockUpdate] = useState(0.0);
-  const [xbtcBlockUpdate, setXbtcBlockUpdate] = useState(0.0);
-  const [usdaBlockUpdate, setUsdaBlockUpdate] = useState(0.0);
-  const [dikoBlockUpdate, setDikoBlockUpdate] = useState(0.0);
-  const [stxBlockAgoUpdate, setStxBlockAgoUpdate] = useState(0.0);
-  const [xbtcBlockAgoUpdate, setXbtcBlockAgoUpdate] = useState(0.0);
-  const [usdaBlockAgoUpdate, setUsdaBlockAgoUpdate] = useState(0.0);
-  const [dikoBlockAgoUpdate, setDikoBlockAgoUpdate] = useState(0.0);
   const [loadingVaults, setLoadingVaults] = useState(true);
-  const [loadingPrices, setLoadingPrices] = useState(true);
-  const [loadingStackingData, setLoadingStackingData] = useState(false);
   const [pendingVaultRewards, setPendingVaultRewards] = useState(0);
-
-  useEffect(() => {
-    const fetchPrices = async () => {
-
-      // Get current block height
-      const client = getRPCClient();
-      const response = await fetch(`${client.url}/v2/info`, { credentials: 'omit' });
-      const data = await response.json();
-      const currentBlock = data['stacks_tip_height'];
-
-      const stxPrice = await getPriceInfo('STX');
-      setStxPrice(stxPrice['last-price'].value);
-      setStxBlockUpdate(stxPrice['last-block'].value);
-      setStxBlockAgoUpdate(currentBlock - stxPrice['last-block'].value)
-
-      const xbtcPrice = await getPriceInfo('xBTC');
-      setXbtcPrice(xbtcPrice['last-price'].value);
-      setXbtcBlockUpdate(xbtcPrice['last-block'].value);
-      setXbtcBlockAgoUpdate(currentBlock - xbtcPrice['last-block'].value)
-
-      const dikoPrice = await getDikoAmmPrice();
-      setDikoPrice(dikoPrice);
-      setDikoBlockUpdate(currentBlock);
-      setDikoBlockAgoUpdate(1)
-
-      const usdaPrice = await getPriceInfo('USDA');
-      setUsdaPrice(usdaPrice['last-price'].value);
-      setUsdaBlockUpdate(usdaPrice['last-block'].value);
-      setUsdaBlockAgoUpdate(currentBlock - usdaPrice['last-block'].value)
-
-      setLoadingStackingData(false);
-      setLoadingPrices(false);
-    };
-
-    setLoadingPrices(true);
-    setLoadingStackingData(true);
-    fetchPrices();
-  }, []);
-
-  useEffect(() => {
-    if (state.currentTxStatus === 'success') {
-      window.location.reload();
-    }
-  }, [state.currentTxStatus]);
 
   useEffect(() => {
     const fetchVault = async (vaultId: number) => {
@@ -170,20 +105,6 @@ export const Mint = () => {
     fetchVaults();
   }, []);
 
-  const addMocknetStx = async () => {
-    const key = '753b7cc01a1a2e86221266a154af739463fce51219d97e4f856cd7200c3bd2a601';
-    const senderKey = createStacksPrivateKey(key);
-    console.log('Adding STX from mocknet address to', address, 'on network', network);
-
-    const transaction = await makeSTXTokenTransfer({
-      recipient: standardPrincipalCV(address || ''),
-      amount: new BN(5000000000),
-      senderKey: privateKeyToString(senderKey),
-      network: network,
-    });
-    await broadcastTransaction(transaction, network);
-  };
-
   const claimPendingRewards = async () => {
     await doContractCall({
       network,
@@ -236,49 +157,33 @@ export const Mint = () => {
                   Stacking Cycle #
                 </dt>
                 <dd className="flex items-baseline justify-between mt-1 md:block lg:flex">
-                  {loadingStackingData === true ? (
-                    <Placeholder className="py-2" width={Placeholder.width.THIRD} />
-                  ) : (
-                    <div className="flex items-baseline text-2xl font-semibold text-indigo-600 dark:text-indigo-100">
-                      {state.cycleNumber}
-                    </div>
-                  )}
+                  <div className="flex items-baseline text-2xl font-semibold text-indigo-600 dark:text-indigo-100">
+                    {state.cycleNumber}
+                  </div>
                 </dd>
               </div>
               <div className="px-4 py-5 sm:p-6">
                 <dt className="text-xs font-semibold text-indigo-800 uppercase dark:text-indigo-200">End date</dt>
                 <dd className="flex items-baseline justify-between mt-1 md:block lg:flex">
-                  {loadingStackingData === true ? (
-                    <Placeholder className="py-2" width={Placeholder.width.THIRD} />
-                  ) : (
-                    <div className="flex items-baseline text-2xl font-semibold text-indigo-600 dark:text-indigo-100">
-                      {state.endDate}
-                    </div>
-                  )}
+                  <div className="flex items-baseline text-2xl font-semibold text-indigo-600 dark:text-indigo-100">
+                    {state.endDate}
+                  </div>
                 </dd>
               </div>
               <div className="px-4 py-5 sm:p-6">
                 <dt className="text-xs font-semibold text-indigo-800 uppercase dark:text-indigo-200">Days in cycle</dt>
                 <dd className="flex items-baseline justify-between mt-1 md:block lg:flex">
-                  {loadingStackingData === true ? (
-                    <Placeholder className="py-2" width={Placeholder.width.THIRD} />
-                  ) : (
-                    <div className="flex items-baseline text-2xl font-semibold text-indigo-600 dark:text-indigo-100">
-                      {state.daysPassed}
-                    </div>
-                  )}
+                  <div className="flex items-baseline text-2xl font-semibold text-indigo-600 dark:text-indigo-100">
+                    {state.daysPassed}
+                  </div>
                 </dd>
               </div>
               <div className="px-4 py-5 sm:p-6">
                 <dt className="text-xs font-semibold text-indigo-800 uppercase dark:text-indigo-200">Days left</dt>
                 <dd className="flex items-baseline justify-between mt-1 md:block lg:flex">
-                  {loadingStackingData === true ? (
-                    <Placeholder className="py-2" width={Placeholder.width.THIRD} />
-                  ) : (
-                    <div className="flex items-baseline text-2xl font-semibold text-indigo-600 dark:text-indigo-100">
-                      {state.daysLeft}
-                    </div>
-                  )}
+                  <div className="flex items-baseline text-2xl font-semibold text-indigo-600 dark:text-indigo-100">
+                    {state.daysLeft}
+                  </div>
                 </dd>
               </div>
             </dl>
@@ -380,7 +285,7 @@ export const Mint = () => {
                     <tr>
                       <th
                         scope="col"
-                        className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase bg-gray-50"
+                        className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase"
                       >
                         <Placeholder color={Placeholder.color.GRAY} />
                       </th>
@@ -404,195 +309,7 @@ export const Mint = () => {
           </div>
         </section>
 
-        <section className="mt-8">
-          <header className="pb-5 border-b border-gray-200 dark:border-zinc-600 sm:flex sm:items-center sm:justify-between">
-            <h3 className="text-lg font-medium leading-6 text-gray-900 font-headings dark:text-zinc-50">Assets</h3>
-            <div className="flex mt-3 sm:mt-0 sm:ml-4">
-              {env == 'mocknet' ? (
-                <div className="flex items-center justify-end">
-                  <span className="px-2 py-1 text-xs text-gray-800 dark:text-zinc-100">Mocknet actions:</span>
-                  <button
-                    type="button"
-                    onClick={() => addMocknetStx()}
-                    className="inline-flex items-center px-3 py-2 text-sm font-normal leading-4 text-indigo-700 bg-indigo-100 border border-transparent rounded-md hover:bg-indigo-200 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-                  >
-                    Get 5000 STX from mocknet
-                  </button>
-                </div>
-              ) : null}
-            </div>
-          </header>
-
-          <div className="flex flex-col mt-4">
-            <div className="-my-2 overflow-x-auto sm:-mx-6 lg:-mx-8">
-              <div className="inline-block min-w-full py-2 align-middle sm:px-6 lg:px-8">
-                <div className="overflow-hidden border border-gray-200 rounded-lg dark:border-zinc-700">
-                  <table className="min-w-full divide-y divide-gray-200 dark:divide-zinc-600">
-                    <thead className="bg-gray-50 dark:bg-zinc-900 dark:bg-opacity-80">
-                      <tr>
-                        <th
-                          scope="col"
-                          className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase dark:text-zinc-400"
-                        >
-                          Asset
-                        </th>
-                        <th
-                          scope="col"
-                          className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase dark:text-zinc-400"
-                        >
-                          Last Oracle Price
-                        </th>
-                        <th
-                          scope="col"
-                          className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase dark:text-zinc-400"
-                        >
-                          Updated Block Height
-                        </th>
-                      </tr>
-                    </thead>
-                    <tbody className="bg-white divide-y divide-gray-200 dark:bg-zinc-900 dark:divide-zinc-600">
-                      <tr className="bg-white dark:bg-zinc-900">
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          <div className="flex items-center">
-                            <div className="w-10 h-10 shrink-0">
-                              <img
-                                className="w-10 h-10 rounded-full"
-                                src={tokenList[2].logo}
-                                alt=""
-                              />
-                            </div>
-                            <div className="ml-4">
-                              <div className="text-sm font-medium text-gray-900 dark:text-zinc-100">STX</div>
-                            </div>
-                          </div>
-                        </td>
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          {loadingPrices ? (
-                            <Placeholder className="py-2" width={Placeholder.width.HALF} />
-                          ) : (
-                            <span>${stxPrice / 1000000}</span>
-                          )}
-                        </td>
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          {loadingPrices ? (
-                            <Placeholder className="py-2" width={Placeholder.width.HALF} />
-                          ) : (
-                            <>
-                            <span>{stxBlockUpdate} </span>
-                            <span className="text-gray-500 dark:text-zinc-300">({stxBlockAgoUpdate} blocks ago)</span>
-                            </>
-                          )}
-                        </td>
-                      </tr>
-
-                      <tr className="bg-white dark:bg-zinc-900">
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          <div className="flex items-center">
-                            <div className="w-10 h-10 shrink-0">
-                              <img
-                                className="w-10 h-10 rounded-full"
-                                src={tokenList[1].logo}
-                                alt=""
-                              />
-                            </div>
-                            <div className="ml-4">
-                              <div className="text-sm font-medium text-gray-900 dark:text-zinc-100">DIKO</div>
-                            </div>
-                          </div>
-                        </td>
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          {loadingPrices ? (
-                            <Placeholder className="py-2" width={Placeholder.width.HALF} />
-                          ) : (
-                            <span>{dikoPrice} USDA</span>
-                          )}
-                        </td>
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          {loadingPrices ? (
-                            <Placeholder className="py-2" width={Placeholder.width.HALF} />
-                          ) : (
-                            <>
-                            <span>{dikoBlockUpdate} </span>
-                            <span className="text-gray-500 dark:text-zinc-300">({dikoBlockAgoUpdate} blocks ago)</span>
-                            </>
-                          )}
-                        </td>
-                      </tr>
-
-                      <tr className="bg-white dark:bg-zinc-900">
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          <div className="flex items-center">
-                            <div className="w-10 h-10 shrink-0">
-                              <img
-                                className="w-10 h-10 rounded-full"
-                                src={tokenList[3].logo}
-                                alt=""
-                              />
-                            </div>
-                            <div className="ml-4">
-                              <div className="text-sm font-medium text-gray-900 dark:text-zinc-100">xBTC</div>
-                            </div>
-                          </div>
-                        </td>
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          {loadingPrices ? (
-                            <Placeholder className="py-2" width={Placeholder.width.HALF} />
-                          ) : (
-                            <span>${xbtcPrice / 1000000}</span>
-                          )}
-                        </td>
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          {loadingPrices ? (
-                            <Placeholder className="py-2" width={Placeholder.width.HALF} />
-                          ) : (
-                            <>
-                            <span>{xbtcBlockUpdate} </span>
-                            <span className="text-gray-500 dark:text-zinc-300">({xbtcBlockAgoUpdate} blocks ago)</span>
-                            </>
-                          )}
-                        </td>
-                      </tr>
-
-                      <tr className="bg-white dark:bg-zinc-900">
-                        <td className="px-6 py-4 text-sm text-gray-500 whitespace-nowrap">
-                          <div className="flex items-center">
-                            <div className="w-10 h-10 shrink-0">
-                              <img
-                                className="w-10 h-10 rounded-full"
-                                src={tokenList[0].logo}
-                                alt=""
-                              />
-                            </div>
-                            <div className="ml-4">
-                              <div className="text-sm font-medium text-gray-900 dark:text-zinc-100">USDA</div>
-                            </div>
-                          </div>
-                        </td>
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          {loadingPrices ? (
-                            <Placeholder className="py-2" width={Placeholder.width.HALF} />
-                          ) : (
-                            <span>${usdaPrice / 1000000}</span>
-                          )}
-                        </td>
-                        <td className="px-6 py-4 text-sm text-gray-900 dark:text-zinc-100 whitespace-nowrap">
-                          {loadingPrices ? (
-                            <Placeholder className="py-2" width={Placeholder.width.HALF} />
-                          ) : (
-                            <>
-                            <span>{usdaBlockUpdate} </span>
-                            <span className="text-gray-500 dark:text-zinc-300">({usdaBlockAgoUpdate} blocks ago)</span>
-                            </>
-                          )}
-                        </td>
-                      </tr>
-                    </tbody>
-                  </table>
-                </div>
-              </div>
-            </div>
-          </div>
-        </section>
+        <Prices />
       </main>
     </div>
   );
